requirement.txt
pandas==2.2.2
numpy==1.26.4
plotly==5.20.0
dash==2.17.0
streamlit==1.34.0
Pillow==10.2.0